Make it modular.
