+++ src/styles/arrow.ts
@@ -1,0 +1,51 @@
+import {
+  ArcStyle,
+  markVolatile,
+  mercatorProjection,
+  VcsApp,
+  VectorLayer,
+} from '@vcmap/core'
+import type { LineString } from 'ol/geom'
+import { Feature } from 'ol'
+
+export function getArrowScratchLayer(
+  app: VcsApp,
+  layerName: string
+): VectorLayer {
+  if (app.layers.hasKey(layerName)) {
+    return app.layers.getByKey(layerName) as VectorLayer
+  }
+
+  const layer = new VectorLayer({
+    name: layerName,
+    projection: mercatorProjection.toJSON(),
+  })
+
+  markVolatile(layer)
+  app.layers.add(layer)
+  layer.activate()
+  return layer
+}
+
+export function updateArrowFeatures(
+  linestrings: LineString[],
+  arrowLayer: VectorLayer
+) {
+  const features: Feature[] = []
+  linestrings.forEach((linestring) => {
+    features.push(new Feature(linestring))
+  })
+  arrowLayer.removeAllFeatures()
+  arrowLayer.addFeatures(features)
+}
+
+export function updateArrowLayerStyle(arrowLayer: VectorLayer, is3D: boolean) {
+  // Update arrow's style
+  let arrowColor = '#000000'
+  if (is3D) {
+    arrowColor = '#FFFFFF'
+  }
+  arrowLayer.setStyle(
+    new ArcStyle({ width: 1.5, arcFactor: 0.25, color: arrowColor })
+  )
+}
